+++ src/Nodes/index.ts
@@ -1,129 +1,29 @@
-export enum CmdSyntaxKind {
-	Unknown,
-	Source,
+import { CmdSyntaxKind, NodeFlag } from "./Enum";
+import {
+	NodeBase,
+	NodeError,
+	StringLiteral,
+	NumberLiteral,
+	InterpolatedStringExpression,
+	BooleanLiteral,
+	CommandName,
 	CommandStatement,
 	InnerExpression,
-	String,
-	Boolean,
-	CommandName,
-	Number,
-	Option,
-	Identifier,
+	CommandSource,
 	OperatorToken,
-	BinaryExpression,
-	InterpolatedString,
-	PrefixToken,
-	PrefixExpression,
-	EndOfStatement,
 	VariableDeclaration,
 	VariableStatement,
-	Invalid,
-}
-
-export const enum NodeFlag {
-	None = 0,
-	NodeHasError = 1 << 16,
-}
+	InvalidNode,
+	BinaryExpression,
+	Node,
+	NodeTypes,
+	ParentNode,
+	VALID_PREFIX_CHARS,
+} from "./NodeTypes";
+export { CmdSyntaxKind, NodeFlag };
 
 export function hasNodeFlag<F extends NodeFlag>(node: Node, flag: F) {
 	return node.flags !== undefined && (node.flags & flag) !== 0;
-}
-
-interface NodeBase {
-	kind: CmdSyntaxKind;
-	parent?: Node;
-	pos?: number;
-	rawText?: string;
-	endPos?: number;
-	flags: NodeFlag;
-}
-
-type OP = "&&" | "|" | "=";
-
-export interface OperatorToken extends NodeBase {
-	operator: OP;
-	kind: CmdSyntaxKind.OperatorToken;
-}
-
-export interface CommandSource extends NodeBase {
-	kind: CmdSyntaxKind.Source;
-	children: Array<Node>;
-}
-
-export interface InterpolatedStringExpression extends NodeBase {
-	kind: CmdSyntaxKind.InterpolatedString;
-	values: Array<StringLiteral | Identifier>;
-}
-
-export interface BinaryExpression extends NodeBase {
-	kind: CmdSyntaxKind.BinaryExpression;
-	left: Node;
-	operator: OperatorToken;
-	right: Node;
-	children: Node[];
-}
-
-export interface InvalidNode extends NodeBase {
-	kind: CmdSyntaxKind.Invalid;
-	nodes?: Node[];
-	message: string;
-}
-
-export interface VariableDeclaration extends NodeBase {
-	kind: CmdSyntaxKind.VariableDeclaration;
-	modifiers?: never;
-	identifier: Identifier;
-	expression:
-		| NumberLiteral
-		| StringLiteral
-		| InterpolatedStringExpression
-		| BooleanLiteral
-		| Identifier
-		| InnerExpression;
-}
-
-export interface VariableStatement extends NodeBase {
-	kind: CmdSyntaxKind.VariableStatement;
-	declaration: VariableDeclaration;
-}
-
-export interface CommandName extends NodeBase {
-	kind: CmdSyntaxKind.CommandName;
-	name: StringLiteral;
-}
-
-export interface StringLiteral extends NodeBase {
-	kind: CmdSyntaxKind.String;
-	quotes?: string;
-	isUnterminated?: boolean;
-	text: string;
-}
-
-export interface BooleanLiteral extends NodeBase {
-	kind: CmdSyntaxKind.Boolean;
-	value: boolean;
-}
-
-export interface NumberLiteral extends NodeBase {
-	kind: CmdSyntaxKind.Number;
-	value: number;
-}
-
-export interface CommandStatement extends NodeBase {
-	kind: CmdSyntaxKind.CommandStatement;
-	command: CommandName;
-	isUnterminated?: boolean;
-	children: Node[];
-}
-
-export interface InnerExpression extends NodeBase {
-	kind: CmdSyntaxKind.InnerExpression;
-	expression: CommandStatement | BinaryExpression | VariableStatement;
-}
-
-export interface NodeError {
-	node: Node;
-	message: string;
 }
 
 export function createNodeError(message: string, node: Node): NodeError {
@@ -351,36 +251,6 @@ export function offsetNodePosition(node: Node, offset: number) {
 	}
 }
 
-/////////////////////////////////////////////////
-// Checks
-/////////////////////////////////////////////////
-export interface NodeTypes {
-	[CmdSyntaxKind.CommandStatement]: CommandStatement;
-	[CmdSyntaxKind.CommandName]: CommandName;
-	[CmdSyntaxKind.String]: StringLiteral;
-	[CmdSyntaxKind.Option]: Option;
-	[CmdSyntaxKind.EndOfStatement]: EndOfStatement;
-	[CmdSyntaxKind.Source]: CommandSource;
-	[CmdSyntaxKind.Identifier]: Identifier;
-	[CmdSyntaxKind.Boolean]: BooleanLiteral;
-	[CmdSyntaxKind.Number]: NumberLiteral;
-	[CmdSyntaxKind.InterpolatedString]: InterpolatedStringExpression;
-	[CmdSyntaxKind.BinaryExpression]: BinaryExpression;
-	[CmdSyntaxKind.OperatorToken]: OperatorToken;
-	[CmdSyntaxKind.PrefixToken]: PrefixToken;
-	[CmdSyntaxKind.PrefixExpression]: PrefixExpression;
-	[CmdSyntaxKind.VariableDeclaration]: VariableDeclaration;
-	[CmdSyntaxKind.VariableStatement]: VariableStatement;
-	[CmdSyntaxKind.Invalid]: InvalidNode;
-	[CmdSyntaxKind.InnerExpression]: InnerExpression;
-}
-
-type NonParentNode<T> = T extends { children: Node[] } ? never : T;
-export type ParentNode = Exclude<Node, NonParentNode<Node>>;
-
-export type NodeKind = keyof NodeTypes;
-export type Node = NodeTypes[keyof NodeTypes];
-
 export function isNode<K extends keyof NodeTypes>(node: Node, type: K): node is NodeTypes[K] {
 	return node !== undefined && node.kind === type;
 }
@@ -399,7 +269,6 @@ export function isParentNode(node: Node): node is ParentNode {
 	return "children" in node;
 }
 
-const VALID_PREFIX_CHARS = ["~", "@", "%", "^", "*", "!"] as const;
 export function isValidPrefixCharacter(input: string): input is typeof VALID_PREFIX_CHARS[number] {
 	return VALID_PREFIX_CHARS.includes(input as typeof VALID_PREFIX_CHARS[number]);
 }
